+++ cuda/src/main/java/org/bytedeco/javacpp/presets/cufft.java
@@ -33,8 +33,7 @@ import org.bytedeco.javacpp.tools.InfoMapper;
  * @author Samuel Audet
  */
 @Properties(inherit = cuda.class, value = {
-    @Platform(value = {"linux-x86_64", "linux-ppc64", "macosx-x86_64", "windows-x86_64"},
-              include = "<cufft.h>", link = "cufft@.7.0")},
+    @Platform(include = "<cufft.h>", link = "cufft@.7.0")},
         target = "org.bytedeco.javacpp.cufft")
 public class cufft implements InfoMapper {
     public void map(InfoMap infoMap) {